Use for uintptr 32/64 bit check

+++ yos/platform.go
@@ -21,10 +21,10 @@ func IsOnLinux() bool {
 
 // IsOn32bitArch indicates whether the application is running on 32-bit architecture.
 func IsOn32bitArch() bool {
-	return (^uint(0) >> 31) == 1
+	return (^uintptr(0) >> 31) == 1
 }
 
 // IsOn64bitArch indicates whether the application is running on 64-bit architecture.
 func IsOn64bitArch() bool {
-	return (^uint(0) >> 63) == 1
+	return (^uintptr(0) >> 63) == 1
 }
